+++ features/bootstrap/context_api.go
@@ -3,6 +3,7 @@ package bootstrap
 import (
 	"github.com/DATA-DOG/godog"
 	"github.com/hellofresh/janus/pkg/api"
+	"github.com/pkg/errors"
 )
 
 // RegisterAPIContext registers godog suite context for handling API related steps
@@ -21,13 +22,13 @@ func (c *apiContext) clearAPI(arg interface{}) {
 	if !c.readOnly {
 		data, err := c.apiRepo.FindAll()
 		if err != nil {
-			panic(err)
+			panic(errors.Wrap(err, "Failed to get all registered route specs"))
 		}
 
 		for _, definition := range data {
 			err := c.apiRepo.Remove(definition.Name)
 			if nil != err {
-				panic(err)
+				panic(errors.Wrapf(err, "Failed to remove route spec \"%s\"", definition.Name))
 			}
 		}
 	}