+++ features/bootstrap/context_request.go
@@ -198,8 +198,8 @@ func (c *requestContext) requestJWTTokenIsNotSet() error {
 }
 
 func (c *requestContext) requestJWTTokenIsValidAdminToken() error {
-	jwtConfig := jwt.NewConfigWithHandlers(c.adminCred)
-	accessToken, err := jwt.IssueAdminToken(jwtConfig.SigningAlgorithm, c.adminCred.Username, jwtConfig.Secret, jwtConfig.Timeout)
+	jwtConfig := jwt.NewGuard(c.adminCred)
+	accessToken, err := jwt.IssueAdminToken(jwtConfig.SigningMethod, c.adminCred.Username, jwtConfig.Timeout)
 	if nil != err {
 		return fmt.Errorf("Failed to issue JWT: %v", err)
 	}