+++ pkg/jwt/parser.go
@@ -1,25 +1,73 @@
 package jwt
 
 import (
+	"crypto/rsa"
+	"crypto/x509"
+	"encoding/pem"
 	"errors"
 	"net/http"
 	"strings"
 
-	jwt "github.com/dgrijalva/jwt-go"
+	"github.com/dgrijalva/jwt-go"
 )
+
+var (
+	// ErrSigningMethodMismatch is the error returned when token is signed with the method other than verified
+	ErrSigningMethodMismatch = errors.New("Signing method mismatch")
+	// ErrFailedToParseToken is the error returned when token is failed to parse and validate against secret and expiration date
+	ErrFailedToParseToken = errors.New("Failed to parse token")
+	// ErrUnsupportedSigningMethod is the error returned when token is signed with unsupported by the library method
+	ErrUnsupportedSigningMethod = errors.New("Unsupported signing method")
+	// ErrInvalidPEMBlock is the error returned for keys expected to be PEM-encoded
+	ErrInvalidPEMBlock = errors.New("Invalid RSA: not PEM-encoded")
+	// ErrNotRSAPublicKey is the error returned for invalid RSA public key
+	ErrNotRSAPublicKey = errors.New("Invalid RSA: expected PUBLIC KEY block type")
+	// ErrBadPublicKey is the error returned for invalid RSA public key
+	ErrBadPublicKey = errors.New("Invalid RSA: failed to assert public key")
+)
+
+// SigningMethod defines signing method algorithm and key
+type SigningMethod struct {
+	// Alg defines JWT signing algorithm. Possible values are: HS256, HS384, HS512, RS256, RS384, RS512
+	Alg string `json:"alg"`
+	Key string `json:"key"`
+}
+
+// ParserConfig configures the way JWT Parser gets and validates token
+type ParserConfig struct {
+	// SigningMethods defines chain of token signature verification algorithm/key pairs.
+	SigningMethods []SigningMethod
+
+	// TokenLookup is a string in the form of "<source>:<name>" that is used
+	// to extract token from the request.
+	// Optional. Default value "header:Authorization".
+	// Possible values:
+	// - "header:<name>"
+	// - "query:<name>"
+	// - "cookie:<name>"
+	TokenLookup string
+}
+
+// NewParserConfig creates a new instance of ParserConfig
+func NewParserConfig(signingMethod ...SigningMethod) ParserConfig {
+	return ParserConfig{
+		SigningMethods: signingMethod,
+		TokenLookup:    "header:Authorization",
+	}
+}
 
 // Parser struct
 type Parser struct {
-	Config Config
+	Config ParserConfig
 }
 
 // NewParser creates a new instance of Parser
-func NewParser(config Config) *Parser {
+func NewParser(config ParserConfig) *Parser {
 	return &Parser{config}
 }
 
 // ParseFromRequest tries to extract and validate token from request.
-// See "Config.TokenLookup" for possible ways to pass token in request.
+// See "Guard.TokenLookup" for possible ways to pass token in request.
 func (jp *Parser) ParseFromRequest(r *http.Request) (*jwt.Token, error) {
 	var token string
 	var err error
@@ -42,20 +90,47 @@ func (jp *Parser) ParseFromRequest(r *http.Request) (*jwt.Token, error) {
 }
 
 // Parse a JWT token and validates it
-func (jp *Parser) Parse(token string) (*jwt.Token, error) {
-	return jwt.Parse(token, func(token *jwt.Token) (interface{}, error) {
-		if jwt.GetSigningMethod(jp.Config.SigningAlgorithm) != token.Method {
-			return nil, errors.New("invalid signing algorithm")
+func (jp *Parser) Parse(tokenString string) (*jwt.Token, error) {
+	for _, method := range jp.Config.SigningMethods {
+		token, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
+			if token.Method.Alg() != method.Alg {
+				return nil, ErrSigningMethodMismatch
+			}
+
+			switch token.Method.(type) {
+			case *jwt.SigningMethodHMAC:
+				return []byte(method.Key), nil
+			case *jwt.SigningMethodRSA:
+				block, _ := pem.Decode([]byte(method.Key))
+				if block == nil {
+					return nil, ErrInvalidPEMBlock
+				}
+				if got, want := block.Type, "PUBLIC KEY"; got != want {
+					return nil, ErrNotRSAPublicKey
+				}
+				pub, err := x509.ParsePKIXPublicKey(block.Bytes)
+				if nil != err {
+					return nil, err
+				}
+
+				if _, ok := pub.(*rsa.PublicKey); !ok {
+					return nil, ErrBadPublicKey
+				}
+
+				return pub, nil
+			default:
+				return nil, ErrUnsupportedSigningMethod
+			}
+		})
+
+		if err != nil {
+			continue
 		}
 
-		return jp.Config.Secret, nil
-	})
-}
+		return token, err
+	}
 
-// GetStandardClaims returns a structured version of Claims Section
-func (jp *Parser) GetStandardClaims(token *jwt.Token) (jwt.StandardClaims, bool) {
-	claims, ok := token.Claims.(jwt.StandardClaims)
-	return claims, ok
+	return nil, ErrFailedToParseToken
 }
 
 // GetMapClaims returns a map version of Claims Section