+++ pkg/plugin/oauth2/middleware_access_rules_test.go
@@ -14,6 +14,8 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+const signingAlg = "HS256"
+
 func TestBlockJWTByCountry(t *testing.T) {
 	secret := "secret"
 
@@ -21,11 +23,10 @@ func TestBlockJWTByCountry(t *testing.T) {
 		{Predicate: "country == 'de'", Action: "deny"},
 	}
 
-	config := jwt.NewConfig(secret)
-	parser := jwt.NewParser(config)
+	parser := jwt.NewParser(jwt.NewParserConfig(jwt.SigningMethod{Alg: signingAlg, Key: secret}))
 
 	mw := NewRevokeRulesMiddleware(parser, revokeRules)
-	token, err := generateToken(secret)
+	token, err := generateToken(signingAlg, secret)
 	require.NoError(t, err)
 
 	w, err := test.Record(
@@ -48,11 +49,10 @@ func TestBlockJWTByUsername(t *testing.T) {
 		{Predicate: "username == 'test@hellofresh.com'", Action: "deny"},
 	}
 
-	config := jwt.NewConfig(secret)
-	parser := jwt.NewParser(config)
+	parser := jwt.NewParser(jwt.NewParserConfig(jwt.SigningMethod{Alg: signingAlg, Key: secret}))
 
 	mw := NewRevokeRulesMiddleware(parser, revokeRules)
-	token, err := generateToken(secret)
+	token, err := generateToken(signingAlg, secret)
 	require.NoError(t, err)
 
 	w, err := test.Record(
@@ -75,11 +75,10 @@ func TestBlockJWTByIssueDate(t *testing.T) {
 		{Predicate: fmt.Sprintf("iat < %d", time.Now().Add(1*time.Hour).Unix()), Action: "deny"},
 	}
 
-	config := jwt.NewConfig(secret)
-	parser := jwt.NewParser(config)
+	parser := jwt.NewParser(jwt.NewParserConfig(jwt.SigningMethod{Alg: signingAlg, Key: secret}))
 
 	mw := NewRevokeRulesMiddleware(parser, revokeRules)
-	token, err := generateToken(secret)
+	token, err := generateToken(signingAlg, secret)
 	require.NoError(t, err)
 
 	w, err := test.Record(
@@ -102,11 +101,10 @@ func TestBlockJWTByCountryAndIssueDate(t *testing.T) {
 		{Predicate: fmt.Sprintf("country == 'de' && iat < %d", time.Now().Add(1*time.Hour).Unix()), Action: "deny"},
 	}
 
-	config := jwt.NewConfig(secret)
-	parser := jwt.NewParser(config)
+	parser := jwt.NewParser(jwt.NewParserConfig(jwt.SigningMethod{Alg: signingAlg, Key: secret}))
 
 	mw := NewRevokeRulesMiddleware(parser, revokeRules)
-	token, err := generateToken(secret)
+	token, err := generateToken(signingAlg, secret)
 	require.NoError(t, err)
 
 	w, err := test.Record(
@@ -122,14 +120,14 @@ func TestBlockJWTByCountryAndIssueDate(t *testing.T) {
 	assert.Equal(t, http.StatusUnauthorized, w.Code)
 }
 
-func generateToken(secret string) (string, error) {
-	token := basejwt.NewWithClaims(basejwt.SigningMethodHS256, basejwt.MapClaims{
+func generateToken(alg, key string) (string, error) {
+	token := basejwt.NewWithClaims(basejwt.GetSigningMethod(alg), basejwt.MapClaims{
 		"country":  "de",
 		"username": "test@hellofresh.com",
 		"iat":      time.Now().Unix(),
 	})
 
-	return token.SignedString([]byte(secret))
+	return token.SignedString([]byte(key))
 }
 
 func TestEmptyAccessRules(t *testing.T) {
@@ -137,8 +135,7 @@ func TestEmptyAccessRules(t *testing.T) {
 
 	revokeRules := []*oauth.AccessRule{}
 
-	config := jwt.NewConfig(secret)
-	parser := jwt.NewParser(config)
+	parser := jwt.NewParser(jwt.NewParserConfig(jwt.SigningMethod{Alg: signingAlg, Key: secret}))
 
 	mw := NewRevokeRulesMiddleware(parser, revokeRules)
 
@@ -157,11 +154,10 @@ func TestWrongJWT(t *testing.T) {
 		{Predicate: fmt.Sprintf("country == 'de' && iat < %d", time.Now().Add(1*time.Hour).Unix()), Action: "deny"},
 	}
 
-	config := jwt.NewConfig("wrong_secret")
-	parser := jwt.NewParser(config)
+	parser := jwt.NewParser(jwt.NewParserConfig(jwt.SigningMethod{Alg: signingAlg, Key: "wrong secret"}))
 
 	mw := NewRevokeRulesMiddleware(parser, revokeRules)
-	token, err := generateToken("secret")
+	token, err := generateToken(signingAlg, "secret")
 	require.NoError(t, err)
 
 	w, err := test.Record(
@@ -184,11 +180,10 @@ func TestWrongRule(t *testing.T) {
 		{Predicate: "country == 'wrong'", Action: "deny"},
 	}
 
-	config := jwt.NewConfig(secret)
-	parser := jwt.NewParser(config)
+	parser := jwt.NewParser(jwt.NewParserConfig(jwt.SigningMethod{Alg: signingAlg, Key: secret}))
 
 	mw := NewRevokeRulesMiddleware(parser, revokeRules)
-	token, err := generateToken(secret)
+	token, err := generateToken(signingAlg, secret)
 	require.NoError(t, err)
 
 	w, err := test.Record(