+++ pkg/plugin/oauth2/setup.go
@@ -38,13 +38,13 @@ func setupOAuth2(route *proxy.Route, p plugin.Params) error {
 		return err
 	}
 
-	secret, err := oauthServer.TokenStrategy.Settings.GetJWTSecret()
+	signingMethods, err := oauthServer.TokenStrategy.Settings.GetJWTSigningMethods()
 	if err != nil {
 		return err
 	}
 
 	route.AddInbound(NewKeyExistsMiddleware(manager))
-	route.AddInbound(NewRevokeRulesMiddleware(jwt.NewParser(jwt.NewConfig(secret)), oauthServer.AccessRules))
+	route.AddInbound(NewRevokeRulesMiddleware(jwt.NewParser(jwt.NewParserConfig(signingMethods...)), oauthServer.AccessRules))
 
 	return nil
 }